Prevent target block attributes from being reused in other blocks
We have a lot of support for textDocument/definition requests for
attributes within a target block. Some of them assumed that these
attributes would only ever be found on a target block so if they were
replicated inside a different block (such as group or variable) then the
textDocument/definition request would still get results back. Additional
checks have now been introduced to ensure that we only respond to those
requests if they were sent for a position inside a target block.

Signed-off-by: Remy Suen <remy.suen@docker.com>