Ignore errors related to looking up images
It may not be possible to get metadata about an image for any wide
variety of reasons (network instability, authentication issues, remote
server outage). Should these errors occur, we should ignore them and not
surface them to the user as it is more likely for these errors to be
outside the user's control. Ideally we should only surface warnings and
errors that the user can fix themselves.

Signed-off-by: Remy Suen <remy.suen@docker.com>