Remove unnecessary diagnostics calculations by debouncing
Originally, every textDocument/didChange notification would result in a
diagnostic calculation inside of a goroutine. However, this is not ideal
as multiple events can come in in a short amount of time when the user
is typing. The user is only concerned about the diagnostics of a file
when they are done editing. Any diagnostics pertaining to the file when
they are still in the middle of typing is not useful as the content of
the document has already changed.

By debouncing the diagnostics calculation, we can improve the overall
performance of the language server by not wasting CPU cycles calculating
something that will simply be discarded.

Signed-off-by: Remy Suen <remy.suen@docker.com>