Recover directly in the glsp handler
We noticed an issue where the language server would crash without any
telemetry being sent. This was because the initial implementation of the
Recover handler was not correct since calling recover() from the
deferred function actually returns nil. The panic handling has been
refactored so that the deferred function will call recover() and then
pass that result to the panic handler.

Signed-off-by: Remy Suen <remy.suen@docker.com>