Use the Dockerfile in the specified context for calculating definitions
When a textDocument/definition request is received for Bake files, we
need to use the referenced Dockerfile for determining what the response
should be. If a block has a context attribute, then we should use that
for looking up the Dockerfile instead of assuming that the Dockerfile is
in the same folder as the Bake file.

Signed-off-by: Remy Suen <remy.suen@docker.com>