Get Bake target tree using Go APIs
It is extremely costly to run Bake as a separate process when the
language client is sending requests to the server non-stop. After
reviewing the Bake code, it seems like we can use the ReadTargets
function to get the same result as running Bake directly. By using this
function, we have now improved the performance of the language server
significantly and it is now more in line with the original lookup code
that was implemented in the beginning of the project.

Signed-off-by: Remy Suen <remy.suen@docker.com>