Update README.md with Compose features and telemetry details
We have added new Compose features that we can call out in the readme.
Since telemetry can be configured on initialization we should also call
that out in that section of the readme to make it clear to help remind
users that we do collect telemetry but it can be disabled as desired.

Signed-off-by: Remy Suen <remy.suen@docker.com>