Create a setting to disable Compose completions
There are other language servers that provide Compose completions. By
introducing an experimental setting to disable Compose completions, we
can provide users with the ability to filter out our completions if
desired. We will come back to this decision in the future as the Compose
completions feature continues to be developed and refined.

Signed-off-by: Remy Suen <remy.suen@docker.com>