Refactor path handling code
As part of the refactoring for the URIs with \\wsl$ as a host, some
functions were replaced and are no longer called as we created
centralized code paths so that URI handling code can all be in one
place. We can now delete some code that is no longer called and has been
completely replaced.

Signed-off-by: Remy Suen <remy.suen@docker.com>